+++ OneBlink.SDK/models/GeneratePDFFromSubmissionDataRequest.cs
@@ -26,5 +26,9 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public string pdfSize
+        {
+            get; set;
+        }
     }
 }